requirements.txt
coreapi==2.0.8
Django==1.10.2
django-cors-headers==1.2.1
django-filter==0.15.2
django-rest-swagger==2.0.6
djangorestframework==3.4.7
itypes==1.1.0
openapi-codec==1.1.3
Pillow==3.4.0
PyJWT==1.4.2
requests==2.11.1
simplejson==3.8.2
uritemplate==3.0.0
